requirements.txt
django
djangorestframework
pillow
psycopg2
requests
rfc3987
